Updated generate_samples for efast frequencies.
Modified generate_samples to read an efast frequency file and produced a
sample file for each parameter. A file corresponds to a row in the
frequency file. If the frequency file is only an array (as opposed to a
matrix), the data is assumed to be fast, not efast, and generate_samples
functions as it previously did.